--- delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/application/DeliveryRepository.java
@@ -1,19 +1,0 @@
-package com.fn.eureka.client.deliveryservice.application;
-
-import java.util.Optional;
-import java.util.UUID;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.fn.eureka.client.deliveryservice.domain.delivery.Delivery;
-
-public interface DeliveryRepository extends JpaRepository<Delivery, UUID> {
-
-	Optional<Delivery> findByDeliveryIdAndIsDeletedIsFalse(UUID deliveryId);
-
-	Page<Delivery> findAllByIsDeletedIsFalse(Pageable pageable);
-
-	boolean existsByOrderIdAndIsDeletedIsFalse(UUID orderId);
-}
