+++ delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/application/dto/delivery/response/DeleteDeliveryResponseDto.java
@@ -1,0 +1,11 @@
+package com.fn.eureka.client.deliveryservice.application.dto.delivery.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class DeleteDeliveryResponseDto {
+}
