+++ delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/application/dto/route/request/UpdateHubToHubRequestDto.java
@@ -1,0 +1,14 @@
+package com.fn.eureka.client.deliveryservice.application.dto.route.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class UpdateHubToHubRequestDto {
+
+	private String departureHubAddress;
+	private String arrivalHubAddress;
+}
