--- delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/application/route/dto/request/UpdateHubToHubRequestDto.java
@@ -1,15 +1,0 @@
-package com.fn.eureka.client.deliveryservice.application.route.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-@AllArgsConstructor
-public class UpdateHubToHubRequestDto {
-
-	private String departureHubAddress;
-	private String arrivalHubAddress;
-}
