--- delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/application/route/dto/response/DeleteHubToHubResponseDto.java
@@ -1,11 +1,0 @@
-package com.fn.eureka.client.deliveryservice.application.route.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-@AllArgsConstructor
-public class DeleteHubToHubResponseDto {
-}
