+++ delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/domain/repository/HubToHubRepository.java
@@ -1,0 +1,22 @@
+package com.fn.eureka.client.deliveryservice.domain.repository;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.fn.eureka.client.deliveryservice.domain.model.route.HubToHub;
+
+public interface HubToHubRepository extends JpaRepository<HubToHub, UUID> {
+
+	Optional<HubToHub> findByHthIdAndIsDeletedIsFalse(UUID hubToHubId);
+
+	boolean existsByDepartureHubAddressAndArrivalHubAddressAndIsDeletedIsFalse(String departureHubAddress, String arrivalHubAddress);
+
+	Page<HubToHub> findAllByIsDeletedIsFalse(Pageable pageable);
+
+	List<HubToHub> findAllByIsDeletedIsFalse();
+}
