+++ delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/domain/service/HubToHubServiceImpl.java
@@ -13,15 +13,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fn.common.global.dto.CommonPageResponse;
 import com.fn.common.global.exception.CustomApiException;
-import com.fn.eureka.client.deliveryservice.application.HubToHubRepository;
-import com.fn.eureka.client.deliveryservice.application.HubToHubService;
-import com.fn.eureka.client.deliveryservice.application.route.dto.request.CreateHubToHubRequestDto;
-import com.fn.eureka.client.deliveryservice.application.route.dto.request.UpdateHubToHubRequestDto;
-import com.fn.eureka.client.deliveryservice.application.route.dto.response.CreateHubToHubResponseDto;
-import com.fn.eureka.client.deliveryservice.application.route.dto.response.GetAllHubToHubResponseDto;
-import com.fn.eureka.client.deliveryservice.application.route.dto.response.GetHubToHubResponseDto;
-import com.fn.eureka.client.deliveryservice.application.route.dto.response.UpdateHubToHubResponseDto;
-import com.fn.eureka.client.deliveryservice.domain.route.HubToHub;
+import com.fn.eureka.client.deliveryservice.domain.repository.HubToHubRepository;
+import com.fn.eureka.client.deliveryservice.application.service.HubToHubService;
+import com.fn.eureka.client.deliveryservice.application.dto.route.request.CreateHubToHubRequestDto;
+import com.fn.eureka.client.deliveryservice.application.dto.route.request.UpdateHubToHubRequestDto;
+import com.fn.eureka.client.deliveryservice.application.dto.route.response.CreateHubToHubResponseDto;
+import com.fn.eureka.client.deliveryservice.application.dto.route.response.GetAllHubToHubResponseDto;
+import com.fn.eureka.client.deliveryservice.application.dto.route.response.GetHubToHubResponseDto;
+import com.fn.eureka.client.deliveryservice.application.dto.route.response.UpdateHubToHubResponseDto;
+import com.fn.eureka.client.deliveryservice.domain.model.route.HubToHub;
 import com.fn.eureka.client.deliveryservice.exception.HubToHubException;
 import com.fn.eureka.client.deliveryservice.infrastructure.NaverMapService;
 import com.fn.eureka.client.deliveryservice.presentation.dto.request.NaverMapDirReqestDto;
